reduce code duplication
Move duplicated code into `module_utils/dokku_utils.py`
