+++ .gitignore
@@ -1,3 +1,4 @@
 ansible-role-requirements.yml
 .idea/
 *swp
+.vscode
